Cleanup part 2  : Amber Core SL and ecs system revised for c++17. with static_cast.

+++ AmberCore_SL/Amber/ECS/Ecs.cpp
@@ -5,7 +5,7 @@
 ECS::~ECS()
 {
 	for(Map<uint32, Array<uint8>>::iterator it = components.begin(); it != components.end(); ++it) {
-		size_t typeSize = BaseECSComponent::getTypeSize(it->first);
+		uint32 typeSize = static_cast<uint32>(BaseECSComponent::getTypeSize(it->first));
 		ECSComponentFreeFunction freefn = BaseECSComponent::getTypeFreeFunction(it->first);
 		for(uint32 i = 0; i < it->second.size(); i += typeSize) {
 			freefn((BaseECSComponent*)&it->second[i]);
@@ -31,7 +31,7 @@ EntityHandle ECS::makeEntity(BaseECSComponent** entityComponents, const uint32* 
 		addComponentInternal(handle, newEntity->second, componentIDs[i], entityComponents[i]);
 	}
 
-	newEntity->first = entities.size();
+	newEntity->first = static_cast<uint32>(entities.size());
 	entities.push_back(newEntity);
 
 	for(uint32 i = 0; i < listeners.size(); i++) {
@@ -95,7 +95,7 @@ void ECS::removeEntity(EntityHandle handle)
 	}
 
 	uint32 destIndex = handleToEntityIndex(handle);
-	uint32 srcIndex = entities.size() - 1;
+	uint32 srcIndex = static_cast<uint32>(entities.size() - 1);
 	delete entities[destIndex];
 	entities[destIndex] = entities[srcIndex];
 	entities[destIndex]->first = destIndex;
@@ -116,7 +116,7 @@ void ECS::deleteComponent(uint32 componentID, uint32 index)
 	Array<uint8>& array = components[componentID];
 	ECSComponentFreeFunction freefn = BaseECSComponent::getTypeFreeFunction(componentID);
 	size_t typeSize = BaseECSComponent::getTypeSize(componentID);
-	uint32 srcIndex = array.size() - typeSize;
+	uint32 srcIndex = static_cast<uint32>(array.size() - typeSize);
 
 	BaseECSComponent* destComponent = (BaseECSComponent*)&array[index];
 	BaseECSComponent* srcComponent = (BaseECSComponent*)&array[srcIndex];
@@ -145,7 +145,7 @@ bool ECS::removeComponentInternal(EntityHandle handle, uint32 componentID)
 	for(uint32 i = 0; i < entityComponents.size(); i++) {
 		if(componentID == entityComponents[i].first) {
 			deleteComponent(entityComponents[i].first, entityComponents[i].second);
-			uint32 srcIndex = entityComponents.size()-1;
+			uint32 srcIndex = static_cast<uint32>(entityComponents.size()-1);
 			uint32 destIndex = i;
 			entityComponents[destIndex] = entityComponents[srcIndex];
 			entityComponents.pop_back();
@@ -172,7 +172,7 @@ void ECS::updateSystems(ECSSystemList& systems, float delta)
 	for(uint32 i = 0; i < systems.size(); i++) {
 		const Array<uint32>& componentTypes = systems[i]->getComponentTypes();
 		if(componentTypes.size() == 1) {
-			size_t typeSize = BaseECSComponent::getTypeSize(componentTypes[0]);
+			uint32 typeSize = static_cast<uint32>(BaseECSComponent::getTypeSize(componentTypes[0]));
 			Array<uint8>& array = components[componentTypes[0]];
 			for(uint32 j = 0; j < array.size(); j += typeSize) {
 				BaseECSComponent* component = (BaseECSComponent*)&array[j];
@@ -192,8 +192,8 @@ uint32 ECS::findLeastCommonComponent(const Array<uint32>& componentTypes, const 
 		if((componentFlags[i] & BaseECSSystem::FLAG_OPTIONAL) != 0) {
 			continue;
 		}
-		size_t typeSize = BaseECSComponent::getTypeSize(componentTypes[i]);
-		uint32 size = components[componentTypes[i]].size()/typeSize;
+		uint32 typeSize = static_cast<uint32>(BaseECSComponent::getTypeSize(componentTypes[i]));
+		uint32 size = static_cast<uint32>(components[componentTypes[i]].size()/typeSize);
 		if(size <= minSize) {
 			minSize = size;
 			minIndex = i;
@@ -209,14 +209,14 @@ void ECS::updateSystemWithMultipleComponents(uint32 index, ECSSystemList& system
 {
 	const Array<uint32>& componentFlags = systems[index]->getComponentFlags();
 
-    componentParam.resize(fmax(componentParam.size(), componentTypes.size()));
-    componentArrays.resize(fmax(componentArrays.size(), componentTypes.size()));
+    componentParam.resize(static_cast<uint64>(fmax(componentParam.size(), componentTypes.size())));
+    componentArrays.resize(static_cast<uint64>(fmax(componentArrays.size(), componentTypes.size())));
 	for(uint32 i = 0; i < componentTypes.size(); i++) {
 		componentArrays[i] = &components[componentTypes[i]];
 	}
 	uint32 minSizeIndex = findLeastCommonComponent(componentTypes, componentFlags);
 
-	size_t typeSize = BaseECSComponent::getTypeSize(componentTypes[minSizeIndex]);
+	uint32 typeSize = static_cast<uint32>(BaseECSComponent::getTypeSize(componentTypes[minSizeIndex]));
 	Array<uint8>& array = *componentArrays[minSizeIndex];
 	for(uint32 i = 0; i < array.size(); i += typeSize) {
 		componentParam[minSizeIndex] = (BaseECSComponent*)&array[i];
